Allow promise chaining from takeScreenshot()
When promises are returned, even inside other promises, they can be
chained as the whole line of promises have to resolve before the top one
does.

In this case, `takeScreenshot()` returns a promise that resolves when
a screenshot has been taken, the browser name has been fetched, and the
screenshot has been written (synchronously).